Finish fractal demo
Add multithread implementation

+++ demos/fractal/float_t.hpp
@@ -2,7 +2,7 @@
 #define TERMOX_DEMOS_FRACTAL_FLOAT_T_HPP
 namespace fractal {
 
-using Float_t = double;
+using Float_t = long double;
 
 }  // namespace fractal
 #endif  // TERMOX_DEMOS_FRACTAL_FLOAT_T_HPP
